Correct removal of tags prefix
Tags are formated as, for example:

class_name.method

So "class_name." must be removed from the tags before returning the list of
suggestions.
But the script was obtaining the length of prefix that needs to be cut from the
wrong variable. Unfortunately, in the test file the length of these variables
was the same and the problem was not detected.