Syntax: Change priority of instance matching regular expression
This priority change guarantees instances only match if no other syntax
matches. The matching regular expression was also updated to avoid a
positive match of the following scenarios:

* priority if (
* else if (

It is possible that other scenarios are not covered. They will be dealt
with as they are identified and included in the test cases.